Import krb5 1.3.2.
git-svn-id: svn+ssh://svn.mit.edu/athena/branches/vendor@20172 728af825-273c-0410-89f9-f7d3b574a069

+++ third/krb5/src/lib/kdb/fetch_mkey.c
@@ -95,7 +95,7 @@ krb5_db_fetch_mkey(context, mname, etype, fromkeyboard, twice, keyfile,
     krb5_error_code retval;
     char password[BUFSIZ];
     krb5_data pwd;
-    int size = sizeof(password);
+    unsigned int size = sizeof(password);
 
     if (fromkeyboard) {
 	krb5_data scratch;
@@ -156,18 +156,20 @@ krb5_db_fetch_mkey(context, mname, etype, fromkeyboard, twice, keyfile,
 	    retval = KRB5_KDB_CANTREAD_STORED;
 	    goto errout;
 	}
-	if (!key->length || key->length < 0) {
+	if (!key->length || ((int) key->length) < 0) {
 	    retval = KRB5_KDB_BADSTORED_MKEY;
 	    goto errout;
 	}
+	
 	if (!(key->contents = (krb5_octet *)malloc(key->length))) {
 	    retval = ENOMEM;
 	    goto errout;
 	}
 	if (fread((krb5_pointer) key->contents,
-		  sizeof(key->contents[0]), key->length, kf) != key->length) {
+		  sizeof(key->contents[0]), key->length, kf) 
+	    != key->length) {
 	    retval = KRB5_KDB_CANTREAD_STORED;
-	    memset(key->contents, 0, key->length);
+	    memset(key->contents, 0,  key->length);
 	    free(key->contents);
 	    key->contents = 0;
 	} else
